Refactor v2 validator context support
